(issue14) Move inter_agent_comm to respective app/helpers folder
